+++ src/main/php/lang/reflect/Routine.class.php
@@ -129,7 +129,9 @@ class Routine implements Value {
     } else if ($t instanceof \ReflectionUnionType) {
       $union= [];
       foreach ($t->getTypes() as $component) {
-        $union[]= Type::resolve($component->getName(), $this->resolve());
+        if ('null' !== ($name= $component->getName())) {
+          $union[]= Type::resolve($name, $this->resolve());
+        }
       }
       return new TypeUnion($union);
     } else {
@@ -170,8 +172,9 @@ class Routine implements Value {
     } else if ($t instanceof \ReflectionUnionType) {
       $union= '';
       foreach ($t->getTypes() as $component) {
-        $name= $component->getName();
-        $union.= '|'.($map[$name] ?? strtr($name, '\\', '.'));
+        if ('null' !== ($name= $component->getName())) {
+          $union.= '|'.($map[$name] ?? strtr($name, '\\', '.'));
+        }
       }
       return substr($union, 1);
     } else {
@@ -203,7 +206,9 @@ class Routine implements Value {
       if ($t instanceof \ReflectionUnionType) {
         $union= [];
         foreach ($t->getTypes() as $component) {
-          $union[]= Type::forName($component->getName(), $this->resolve());
+          if ('null' !== ($name= $component->getName())) {
+            $union[]= Type::resolve($name, $this->resolve());
+          }
         }
         return new TypeUnion($union);
       } else {